Completed Entity Details

+++ v2/Log.Realty/app/src/main/java/ug/karuhanga/logrealty/Models/House.java
@@ -74,7 +74,7 @@ public class House extends Record {
         return rent;
     }
 
-    public void setRent(int rent) {
+    public void setRent(long rent) {
         this.rent = rent;
     }
